moved models into seperate module
